refactor(completion): Move suggest test to its own file
